Add unexported precompile() function to facilitate benchmarking
`DataFrames.precompile()` precompiles all methods that are used when running tests.
This is useful to run benchmarks like H2Oai where the time of the first run should
include the time needed to specialize on a particular operation on a particular
dataset, but not the time needed for the more general compilation of common parts.
It could also be useful for users who would like to generate a precompiled image.

To avoid making the package slower to precompile or load, the precompilation code
is only included when actually used.